+++ google/cloud/storage/bucket_metadata.cc
@@ -26,44 +26,6 @@ namespace google {
 namespace cloud {
 namespace storage {
 inline namespace STORAGE_CLIENT_NS {
-namespace {
-CorsEntry ParseCors(internal::nl::json const& json) {
-  auto parse_string_list = [](internal::nl::json const& json,
-                              char const* field_name) {
-    std::vector<std::string> list;
-    if (json.count(field_name) != 0) {
-      for (auto const& kv : json[field_name].items()) {
-        list.emplace_back(kv.value().get<std::string>());
-      }
-    }
-    return list;
-  };
-  CorsEntry result;
-  if (json.count("maxAgeSeconds") != 0) {
-    result.max_age_seconds = internal::ParseLongField(json, "maxAgeSeconds");
-  }
-  result.method = parse_string_list(json, "method");
-  result.origin = parse_string_list(json, "origin");
-  result.response_header = parse_string_list(json, "responseHeader");
-  return result;
-}
-
-void SetIfNotEmpty(internal::nl::json& json, char const* key,
-                   std::string const& value) {
-  if (value.empty()) {
-    return;
-  }
-  json[key] = value;
-}
-
-BucketPolicyOnly ParseBucketOnlyPolicy(internal::nl::json const& json) {
-  BucketPolicyOnly result;
-  result.enabled = internal::ParseBoolField(json, "enabled");
-  result.locked_time = internal::ParseTimestampField(json, "lockedTime");
-  return result;
-}
-
-}  // namespace
 
 std::ostream& operator<<(std::ostream& os, CorsEntry const& rhs) {
   auto join = [](char const* sep, std::vector<std::string> const& list) {
@@ -113,268 +75,6 @@ std::ostream& operator<<(std::ostream& os, BucketRetentionPolicy const& rhs) {
             << rhs.retention_period.count() << "s, effective_time="
             << internal::FormatRfc3339(rhs.effective_time)
             << ", locked=" << rhs.is_locked << "}";
-}
-
-StatusOr<BucketMetadata> BucketMetadata::ParseFromJson(
-    internal::nl::json const& json) {
-  if (!json.is_object()) {
-    return Status(StatusCode::kInvalidArgument, __func__);
-  }
-  BucketMetadata result{};
-  auto status = CommonMetadata<BucketMetadata>::ParseFromJson(result, json);
-  if (!status.ok()) {
-    return status;
-  }
-
-  if (json.count("acl") != 0) {
-    for (auto const& kv : json["acl"].items()) {
-      auto parsed = internal::BucketAccessControlParser::FromJson(kv.value());
-      if (!parsed.ok()) {
-        return std::move(parsed).status();
-      }
-      result.acl_.emplace_back(std::move(*parsed));
-    }
-  }
-  if (json.count("billing") != 0) {
-    auto billing = json["billing"];
-    BucketBilling b;
-    b.requester_pays = internal::ParseBoolField(billing, "requesterPays");
-    result.billing_ = b;
-  }
-  if (json.count("cors") != 0) {
-    for (auto const& kv : json["cors"].items()) {
-      result.cors_.emplace_back(ParseCors(kv.value()));
-    }
-  }
-  if (json.count("defaultEventBasedHold") != 0) {
-    result.default_event_based_hold_ =
-        json.value("defaultEventBasedHold", false);
-  }
-  if (json.count("defaultObjectAcl") != 0) {
-    for (auto const& kv : json["defaultObjectAcl"].items()) {
-      auto parsed = internal::ObjectAccessControlParser::FromJson(kv.value());
-      if (!parsed.ok()) {
-        return std::move(parsed).status();
-      }
-      result.default_acl_.emplace_back(std::move(*parsed));
-    }
-  }
-  if (json.count("encryption") != 0) {
-    BucketEncryption e;
-    e.default_kms_key_name = json["encryption"].value("defaultKmsKeyName", "");
-    result.encryption_ = std::move(e);
-  }
-
-  if (json.count("iamConfiguration") != 0) {
-    BucketIamConfiguration c;
-    auto config = json["iamConfiguration"];
-    if (config.count("bucketPolicyOnly") != 0) {
-      c.bucket_policy_only = ParseBucketOnlyPolicy(config["bucketPolicyOnly"]);
-    }
-    result.iam_configuration_ = c;
-  }
-
-  if (json.count("lifecycle") != 0) {
-    auto lifecycle = json["lifecycle"];
-    BucketLifecycle value;
-    if (lifecycle.count("rule") != 0) {
-      for (auto const& kv : lifecycle["rule"].items()) {
-        auto parsed = internal::LifecycleRuleParser::FromJson(kv.value());
-        if (!parsed.ok()) {
-          return std::move(parsed).status();
-        }
-        value.rule.emplace_back(std::move(*parsed));
-      }
-    }
-    result.lifecycle_ = std::move(value);
-  }
-
-  result.location_ = json.value("location", "");
-
-  if (json.count("logging") != 0) {
-    auto logging = json["logging"];
-    BucketLogging l;
-    l.log_bucket = logging.value("logBucket", "");
-    l.log_object_prefix = logging.value("logObjectPrefix", "");
-    result.logging_ = std::move(l);
-  }
-  result.project_number_ = internal::ParseLongField(json, "projectNumber");
-  if (json.count("labels") > 0) {
-    for (auto const& kv : json["labels"].items()) {
-      result.labels_.emplace(kv.key(), kv.value().get<std::string>());
-    }
-  }
-
-  if (json.count("retentionPolicy") != 0) {
-    auto retention_policy = json["retentionPolicy"];
-    BucketRetentionPolicy r;
-    r.retention_period = std::chrono::seconds(
-        internal::ParseLongField(retention_policy, "retentionPeriod"));
-    r.effective_time =
-        internal::ParseTimestampField(retention_policy, "effectiveTime");
-    r.is_locked = internal::ParseBoolField(retention_policy, "isLocked");
-    result.retention_policy_ = r;
-  }
-
-  if (json.count("versioning") != 0) {
-    auto versioning = json["versioning"];
-    if (versioning.count("enabled") != 0) {
-      BucketVersioning v{internal::ParseBoolField(versioning, "enabled")};
-      result.versioning_ = v;
-    }
-  }
-
-  if (json.count("website") != 0) {
-    auto website = json["website"];
-    BucketWebsite w;
-    w.main_page_suffix = website.value("mainPageSuffix", "");
-    w.not_found_page = website.value("notFoundPage", "");
-    result.website_ = std::move(w);
-  }
-  return result;
-}
-
-StatusOr<BucketMetadata> BucketMetadata::ParseFromString(
-    std::string const& payload) {
-  auto json = storage::internal::nl::json::parse(payload, nullptr, false);
-  return ParseFromJson(json);
-}
-
-std::string BucketMetadata::ToJsonString() const {
-  using internal::nl::json;
-  json metadata_as_json;
-  if (!acl().empty()) {
-    for (BucketAccessControl const& a : acl()) {
-      json entry;
-      SetIfNotEmpty(entry, "entity", a.entity());
-      SetIfNotEmpty(entry, "role", a.role());
-      metadata_as_json["acl"].emplace_back(std::move(entry));
-    }
-  }
-
-  if (!cors().empty()) {
-    for (CorsEntry const& v : cors()) {
-      json cors_as_json;
-      if (v.max_age_seconds.has_value()) {
-        cors_as_json["maxAgeSeconds"] = *v.max_age_seconds;
-      }
-      if (!v.method.empty()) {
-        cors_as_json["method"] = v.method;
-      }
-      if (!v.origin.empty()) {
-        cors_as_json["origin"] = v.origin;
-      }
-      if (!v.response_header.empty()) {
-        cors_as_json["responseHeader"] = v.response_header;
-      }
-      metadata_as_json["cors"].emplace_back(std::move(cors_as_json));
-    }
-  }
-
-  if (has_billing()) {
-    json b{
-        {"requesterPays", billing().requester_pays},
-    };
-    metadata_as_json["billing"] = std::move(b);
-  }
-
-  metadata_as_json["defaultEventBasedHold"] = default_event_based_hold();
-
-  if (!default_acl().empty()) {
-    for (ObjectAccessControl const& a : default_acl()) {
-      json entry;
-      SetIfNotEmpty(entry, "entity", a.entity());
-      SetIfNotEmpty(entry, "role", a.role());
-      metadata_as_json["defaultObjectAcl"].emplace_back(std::move(entry));
-    }
-  }
-
-  if (has_encryption()) {
-    json e;
-    SetIfNotEmpty(e, "defaultKmsKeyName", encryption().default_kms_key_name);
-    metadata_as_json["encryption"] = std::move(e);
-  }
-
-  if (has_iam_configuration()) {
-    json c;
-    if (iam_configuration().bucket_policy_only.has_value()) {
-      json bpo;
-      bpo["enabled"] = iam_configuration().bucket_policy_only->enabled;
-      // The lockedTime field is not mutable and should not be set by the client
-      // the server will provide a value.
-      c["bucketPolicyOnly"] = std::move(bpo);
-    }
-    metadata_as_json["iamConfiguration"] = std::move(c);
-  }
-
-  if (!labels_.empty()) {
-    json labels_as_json;
-    for (auto const& kv : labels_) {
-      labels_as_json[kv.first] = kv.second;
-    }
-    metadata_as_json["labels"] = std::move(labels_as_json);
-  }
-
-  if (has_lifecycle()) {
-    json rule;
-    for (LifecycleRule const& v : lifecycle().rule) {
-      json condition;
-      auto const& c = v.condition();
-      if (c.age) {
-        condition["age"] = *c.age;
-      }
-      if (c.created_before.has_value()) {
-        condition["createdBefore"] = internal::FormatRfc3339(*c.created_before);
-      }
-      if (c.is_live) {
-        condition["isLive"] = *c.is_live;
-      }
-      if (c.matches_storage_class) {
-        condition["matchesStorageClass"] = *c.matches_storage_class;
-      }
-      if (c.num_newer_versions) {
-        condition["numNewerVersions"] = *c.num_newer_versions;
-      }
-      json action{{"type", v.action().type}};
-      if (!v.action().storage_class.empty()) {
-        action["storageClass"] = v.action().storage_class;
-      }
-      rule.emplace_back(json{{"condition", std::move(condition)},
-                             {"action", std::move(action)}});
-    }
-    metadata_as_json["lifecycle"] = json{{"rule", std::move(rule)}};
-  }
-
-  SetIfNotEmpty(metadata_as_json, "location", location());
-
-  if (has_logging()) {
-    json l;
-    SetIfNotEmpty(l, "logBucket", logging().log_bucket);
-    SetIfNotEmpty(l, "logObjectPrefix", logging().log_object_prefix);
-    metadata_as_json["logging"] = std::move(l);
-  }
-
-  SetIfNotEmpty(metadata_as_json, "name", name());
-
-  if (has_retention_policy()) {
-    json r{{"retentionPeriod", retention_policy().retention_period.count()}};
-    metadata_as_json["retentionPolicy"] = std::move(r);
-  }
-
-  SetIfNotEmpty(metadata_as_json, "storageClass", storage_class());
-
-  if (versioning().has_value()) {
-    metadata_as_json["versioning"] = json{{"enabled", versioning()->enabled}};
-  }
-
-  if (has_website()) {
-    json w;
-    SetIfNotEmpty(w, "mainPageSuffix", website().main_page_suffix);
-    SetIfNotEmpty(w, "notFoundPage", website().not_found_page);
-    metadata_as_json["website"] = std::move(w);
-  }
-
-  return metadata_as_json.dump();
 }
 
 bool BucketMetadata::operator==(BucketMetadata const& rhs) const {
